+++ test/spec/Feature/Query/QuerySpec.hs
@@ -939,12 +939,12 @@ spec actualPgVersion = do
         }
 
   describe "Accept headers" $ do
-    it "should respond an unknown accept type with 415" $
+    it "should respond an unknown accept type with 406" $
       request methodGet "/simple_pk"
               (acceptHdrs "text/unknowntype") ""
         `shouldRespondWith`
         [json|{"message":"None of these media types are available: text/unknowntype","code":"PGRST107","details":null,"hint":null}|]
-        { matchStatus  = 415
+        { matchStatus  = 406
         , matchHeaders = [matchContentTypeJson]
         }